Smart formatting in Event Log tab
