+++ neo-gui/UI/DeployContractDialog.cs
@@ -20,7 +20,9 @@ namespace Neo.UI
             byte[] script = textBox8.Text.HexToBytes();
             byte[] parameter_list = textBox6.Text.HexToBytes();
             ContractParameterType return_type = textBox7.Text.HexToBytes().Select(p => (ContractParameterType?)p).FirstOrDefault() ?? ContractParameterType.Void;
-            bool need_storage = checkBox1.Checked;
+            ContractPropertyState properties = ContractPropertyState.NoProperty;
+            if (checkBox1.Checked) properties |= ContractPropertyState.HasStorage;
+            if (checkBox2.Checked) properties |= ContractPropertyState.HasDynamicInvoke;
             string name = textBox1.Text;
             string version = textBox2.Text;
             string author = textBox3.Text;
@@ -31,7 +33,7 @@ namespace Neo.UI
 
             using (ScriptBuilder sb = new ScriptBuilder())
             {
-                sb.EmitSysCall("Neo.Contract.Create", script, parameter_list, return_type, need_storage, name, version, author, email, description);
+                sb.EmitSysCall("Neo.Contract.Create", script, parameter_list, return_type, properties, name, version, author, email, description);
                 return new InvocationTransaction
                 {
                     Script = sb.ToArray()
