+++ neo-gui/UI/DeployContractDialog.zh-Hans.resx
@@ -119,76 +119,76 @@
   </resheader>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
   <data name="textBox5.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 138</value>
+    <value>70, 122</value>
   </data>
   <data name="textBox5.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 84</value>
+    <value>444, 75</value>
   </data>
   <data name="label5.Location" type="System.Drawing.Point, System.Drawing">
-    <value>30, 141</value>
+    <value>30, 125</value>
   </data>
   <data name="label5.Size" type="System.Drawing.Size, System.Drawing">
-    <value>35, 17</value>
+    <value>34, 15</value>
   </data>
   <data name="label5.Text" xml:space="preserve">
     <value>说明:</value>
   </data>
   <data name="textBox4.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 109</value>
+    <value>70, 96</value>
   </data>
   <data name="textBox4.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label4.Location" type="System.Drawing.Point, System.Drawing">
-    <value>6, 112</value>
+    <value>6, 99</value>
   </data>
   <data name="label4.Size" type="System.Drawing.Size, System.Drawing">
-    <value>59, 17</value>
+    <value>58, 15</value>
   </data>
   <data name="label4.Text" xml:space="preserve">
     <value>电子邮件:</value>
   </data>
   <data name="textBox3.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 80</value>
+    <value>70, 71</value>
   </data>
   <data name="textBox3.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label3.Location" type="System.Drawing.Point, System.Drawing">
-    <value>30, 83</value>
+    <value>30, 74</value>
   </data>
   <data name="label3.Size" type="System.Drawing.Size, System.Drawing">
-    <value>35, 17</value>
+    <value>34, 15</value>
   </data>
   <data name="label3.Text" xml:space="preserve">
     <value>作者:</value>
   </data>
   <data name="textBox2.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 51</value>
+    <value>70, 45</value>
   </data>
   <data name="textBox2.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label2.Location" type="System.Drawing.Point, System.Drawing">
-    <value>30, 54</value>
+    <value>30, 48</value>
   </data>
   <data name="label2.Size" type="System.Drawing.Size, System.Drawing">
-    <value>35, 17</value>
+    <value>34, 15</value>
   </data>
   <data name="label2.Text" xml:space="preserve">
     <value>版本:</value>
   </data>
   <data name="textBox1.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 22</value>
+    <value>70, 19</value>
   </data>
   <data name="textBox1.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label1.Location" type="System.Drawing.Point, System.Drawing">
-    <value>30, 25</value>
+    <value>30, 22</value>
   </data>
   <data name="label1.Size" type="System.Drawing.Size, System.Drawing">
-    <value>35, 17</value>
+    <value>34, 15</value>
   </data>
   <data name="label1.Text" xml:space="preserve">
     <value>名称:</value>
@@ -197,28 +197,28 @@
     <value>信息</value>
   </data>
   <data name="textBox7.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 51</value>
+    <value>70, 45</value>
   </data>
   <data name="textBox7.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label7.Location" type="System.Drawing.Point, System.Drawing">
-    <value>18, 54</value>
+    <value>18, 48</value>
   </data>
   <data name="label7.Size" type="System.Drawing.Size, System.Drawing">
-    <value>47, 17</value>
+    <value>46, 15</value>
   </data>
   <data name="label7.Text" xml:space="preserve">
     <value>返回值:</value>
   </data>
   <data name="textBox6.Location" type="System.Drawing.Point, System.Drawing">
-    <value>71, 22</value>
+    <value>70, 19</value>
   </data>
   <data name="textBox6.Size" type="System.Drawing.Size, System.Drawing">
-    <value>443, 23</value>
+    <value>444, 23</value>
   </data>
   <data name="label6.Size" type="System.Drawing.Size, System.Drawing">
-    <value>59, 17</value>
+    <value>58, 15</value>
   </data>
   <data name="label6.Text" xml:space="preserve">
     <value>参数列表:</value>
@@ -226,11 +226,17 @@
   <data name="groupBox2.Text" xml:space="preserve">
     <value>元数据</value>
   </data>
+  <data name="checkBox2.Size" type="System.Drawing.Size, System.Drawing">
+    <value>98, 19</value>
+  </data>
+  <data name="checkBox2.Text" xml:space="preserve">
+    <value>需要动态调用</value>
+  </data>
   <data name="button1.Text" xml:space="preserve">
     <value>加载</value>
   </data>
   <data name="checkBox1.Size" type="System.Drawing.Size, System.Drawing">
-    <value>111, 21</value>
+    <value>110, 19</value>
   </data>
   <data name="checkBox1.Text" xml:space="preserve">
     <value>需要创建存储区</value>
