+++ neo-gui/UI/ElectionDialog.cs
@@ -1,7 +1,6 @@
 ﻿using Neo.Core;
 using Neo.Cryptography.ECC;
-using Neo.SmartContract;
-using Neo.VM;
+using Neo.IO;
 using System;
 using System.Linq;
 using System.Windows.Forms;
@@ -15,25 +14,23 @@ namespace Neo.UI
             InitializeComponent();
         }
 
-        public InvocationTransaction GetTransaction()
+        public StateTransaction GetTransaction()
         {
             ECPoint pubkey = (ECPoint)comboBox1.SelectedItem;
-            using (ScriptBuilder sb = new ScriptBuilder())
+            return Program.CurrentWallet.MakeTransaction(new StateTransaction
             {
-                sb.EmitSysCall("Neo.Validator.Register", pubkey);
-                return new InvocationTransaction
+                Version = 0,
+                Descriptors = new[]
                 {
-                    Attributes = new[]
+                    new StateDescriptor
                     {
-                        new TransactionAttribute
-                        {
-                            Usage = TransactionAttributeUsage.Script,
-                            Data = Contract.CreateSignatureRedeemScript(pubkey).ToScriptHash().ToArray()
-                        }
-                    },
-                    Script = sb.ToArray()
-                };
-            }
+                        Type = StateType.Validator,
+                        Key = pubkey.ToArray(),
+                        Field = "Registered",
+                        Value = BitConverter.GetBytes(true)
+                    }
+                }
+            });
         }
 
         private void ElectionDialog_Load(object sender, EventArgs e)
@@ -43,7 +40,26 @@ namespace Neo.UI
 
         private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
         {
-            button1.Enabled = comboBox1.SelectedIndex >= 0;
+            if (comboBox1.SelectedIndex >= 0)
+            {
+                button1.Enabled = true;
+                ECPoint pubkey = (ECPoint)comboBox1.SelectedItem;
+                StateTransaction tx = new StateTransaction
+                {
+                    Version = 0,
+                    Descriptors = new[]
+                    {
+                        new StateDescriptor
+                        {
+                            Type = StateType.Validator,
+                            Key = pubkey.ToArray(),
+                            Field = "Registered",
+                            Value = BitConverter.GetBytes(true)
+                        }
+                    }
+                };
+                label3.Text = $"{tx.SystemFee} gas";
+            }
         }
     }
 }
