+++ neo-gui/UI/ElectionDialog.es-ES.resx
@@ -117,19 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
-  <data name="label1.Location" type="System.Drawing.Point, System.Drawing">
-    <value>64, 12</value>
-  </data>
-  <data name="label1.Size" type="System.Drawing.Size, System.Drawing">
-    <value>63, 17</value>
-  </data>
-  <data name="label1.Text" xml:space="preserve">
-    <value>Contable:</value>
-  </data>
-  <data name="button1.Text" xml:space="preserve">
-    <value>Aceptar</value>
-  </data>
   <data name="$this.Text" xml:space="preserve">
     <value>Votación</value>
   </data>
