+++ neo-gui/packages.config
@@ -41,7 +41,7 @@
   <package id="Microsoft.Extensions.Primitives" version="2.0.0" targetFramework="net47" />
   <package id="Microsoft.Net.Http.Headers" version="2.0.0" targetFramework="net47" />
   <package id="Microsoft.NETCore.Platforms" version="1.0.1" targetFramework="net47" />
-  <package id="Neo" version="2.5.2" targetFramework="net47" />
+  <package id="Neo" version="2.6.0" targetFramework="net47" />
   <package id="Neo.VM" version="2.0.4" targetFramework="net47" />
   <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net47" />
   <package id="Remotion.Linq" version="2.1.1" targetFramework="net47" />
